Move SettingsFragment to fragments package
